Add Mastro to the benchmark
This repo is a fork of the Eleventy benchmark discussed
on https://www.zachleat.com/web/build-benchmark/.

The numbers on https://mastrojs.github.io were tested on my
MacBook Air M2 with 24GB RAM, on macOS 15.5.
Each run was repeated 3 times and the lowest time was selected.
Interestingly, the Mastro and Eleventy runs for >2000 files
seem to be dominated by directory creation (fs.mkdir).